requirements.txt
langchain_community==0.3.27
python-dotenv==1.1.1
uvicorn==0.35.0
fastapi==0.116.1
aiofiles==24.1.0
langgraph==0.5.3
langgraph-supervisor==0.0.27
langchain-openai==0.3.22
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41